order/requirements.txt
fastapi>=0
starlette>=0
pydantic>=0
flake8>=0
httpx==0.23.0
uvicorn>=0
psycopg2>=0
python-dotenv>=0
asyncpg>=0
sqlalchemy[asyncio]>=0
